Feature/유저 소감 작성하기 비즈니스 로직 (#600)
* fix: BASE_URL이 잘못 들어가는 오류

* fix: change Timber type error to info

* feat: timer 속도 빠르게 변경하고, 서버에서 내려온 값으로 타이머 설정 (#591)

* fix: change time type Int to Double

* fix: 유저 소감 작성하기 UI 디테일 잡기

* feat: update QuizInfoResponse

* style: lint style 적용

* feat: PostQuizReaction 비즈니스 로직 작업

* feat: 이전 기록보다 더 높은 기록일 경우에만 reaction 받기

---------

+++ data/src/main/kotlin/team/duckie/app/android/data/quiz/model/PatchQuizBody.kt
@@ -15,7 +15,7 @@ data class PatchQuizBody(
     @JsonProperty("correctProblemCount")
     val correctProblemCount: Int? = null,
     @JsonProperty("time")
-    val time: Int? = null,
+    val time: Double? = null,
     @JsonProperty("requirementAnswer")
     val requirementAnswer: String? = null,
     @JsonProperty("wrongAnswer")
